ocl: do not use double in kernels
Gen12LP does not support double precision.

+++ src/ocl/gen9_gemm_nocopy_scale_x8x8s32.cl
@@ -63,11 +63,11 @@ kernel void gen9_gemm_scale_x8x8s32(global int *cc, global int *c, char trc,
         if (m > 0) {
             if (!alpha_is_zero) {
                 c[offset_c + 0]
-                        = (int)((double)alpha * cc[offset_cc + 0]
-                                  + (double)beta * c[offset_c + 0] + (0.0))
+                        = (int)((float)alpha * cc[offset_cc + 0]
+                                  + (float)beta * c[offset_c + 0] + (0.0))
                         + co[offset_co + offset_x];
             } else {
-                c[offset_c + 0] = (int)((double)beta * c[offset_c + 0] + (0.0))
+                c[offset_c + 0] = (int)((float)beta * c[offset_c + 0] + (0.0))
                         + co[offset_co + offset_x];
             }
             if (trc == 'C') { offset_x++; }
@@ -76,11 +76,11 @@ kernel void gen9_gemm_scale_x8x8s32(global int *cc, global int *c, char trc,
         if (m > 1) {
             if (!alpha_is_zero) {
                 c[offset_c + 1]
-                        = (int)((double)alpha * cc[offset_cc + 1]
-                                  + (double)beta * c[offset_c + 1] + (0.0))
+                        = (int)((float)alpha * cc[offset_cc + 1]
+                                  + (float)beta * c[offset_c + 1] + (0.0))
                         + co[offset_co + offset_x];
             } else {
-                c[offset_c + 1] = (int)((double)beta * c[offset_c + 1] + (0.0))
+                c[offset_c + 1] = (int)((float)beta * c[offset_c + 1] + (0.0))
                         + co[offset_co + offset_x];
             }
             if (apply_eltwise) UPDATE_ELEM_C(offset_c + 1);
